requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.13.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
dateparser==1.2.0
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.9
dirtyjson==1.0.8
distro==1.9.0
docling==2.29.0
docling-core==2.26.0
docling-ibm-models==3.4.1
docling-parse==4.0.1
easyocr==1.7.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
gotrue==2.12.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-ws==0.7.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
latex2mathml==3.77.0
lazy_loader==0.4
llama-index-core==0.12.10.post1
llama-parse==0.5.19
llvmlite==0.43.0
lmstudio==1.1.0
lxml==5.3.2
markdown-it-py==3.0.0
marko==2.1.3
MarkupSafe==3.0.2
marshmallow==3.24.0
mdurl==0.1.2
more-itertools==10.5.0
mpire==2.10.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.59.3
openai-whisper @ git+https://github.com/openai/whisper.git@90db0de1896c23cbfaf0c58bc2d30665f709f170
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.2.1
pyclipper==1.3.0.post6
pydantic==2.10.4
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
PyMuPDF==1.25.1
pymupdf4llm==0.0.17
PyPDF2==3.0.1
pypdfium2==4.30.1
pytest==8.3.5
pytest-mock==3.14.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
realtime==2.4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
semchunk==2.2.2
setuptools==75.6.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tifffile==2025.3.30
tiktoken==0.8.0
tokenizers==0.21.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.51.2
triton==3.1.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
websockets==14.2
wrapt==1.17.0
wsproto==1.2.0
XlsxWriter==3.2.2
yarl==1.18.3
